pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler==3.10.4",
    "boto3==1.34.162",
    "fastapi==0.115.0",
    "httpx==0.27.0",
    "jinja2==3.1.4",
    "orjson==3.10.7",
    "psycopg[binary]==3.2.1",
    "pydantic-settings>=2.10.1",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "sqlalchemy==2.0.32",
    "uvicorn[standard]==0.30.6",
    "psycopg2-binary>=2.9.10",
    "pillow==10.4.0",
    "scikit-learn==1.5.1",
    "opencv-python-headless==4.10.0.84",
    "imagehash==4.3.1",
    "feedparser>=6.0.11",
    "requests>=2.32.4",
    "numpy>=2.3.2",
    "prometheus-client==0.20.0",
    "tenacity==9.0.0",
    "filelock==3.15.4",
    "praw==7.7.1",
    "google-search-results==2.4.2",
    "openai>=1.99.9",
    "stripe>=12.4.0",
    "itsdangerous>=2.2.0",
]
